requirements.txt
uvloop~=0.19
aiosqlite~=0.19
asyncstdlib~=3.10
discord.py[speed]~=2.3
chromadb-client==0.4.19.dev0
opentelemetry-api>=1.2.0
opentelemetry-sdk>=1.2.0
opentelemetry-exporter-otlp-proto-grpc>=1.2.0
python-dotenv
